move WebGL context to a separate WebGL service independent of the view component, to prevent issues when changing layouts
