+++ src/components/ImageView/RasterView/RasterViewComponent.tsx
@@ -2,13 +2,9 @@ import * as React from "react";
 import {observer} from "mobx-react";
 import {FrameStore, OverlayStore, PreferenceStore, RasterRenderType} from "stores";
 import {FrameView, Point2D, TileCoordinate} from "models";
-import {RasterTile, TEXTURE_SIZE, TILE_SIZE, TileService} from "services/TileService";
-import {GetRequiredTiles, getShaderProgram, GL, LayerToMip, loadFP32Texture, loadImageTexture, hexStringToRgba, add2D, scale2D} from "utilities";
+import {RasterTile, TILE_SIZE, TileService} from "services/TileService";
+import {GetRequiredTiles, GL, LayerToMip, loadFP32Texture, hexStringToRgba, add2D, scale2D} from "utilities";
 import "./RasterViewComponent.css";
-import allMaps from "static/allmaps.png";
-
-const vertexShader = require("!raw-loader!./GLSL/vertex_shader.glsl");
-const pixelShader = require("!raw-loader!./GLSL/pixel_shader_float_rgb.glsl");
 
 export class RasterViewComponentProps {
     preference: PreferenceStore;
@@ -18,141 +14,43 @@ export class RasterViewComponentProps {
     docked: boolean;
 }
 
-interface ShaderUniforms {
-    MinVal: WebGLUniformLocation;
-    MaxVal: WebGLUniformLocation;
-    Bias: WebGLUniformLocation;
-    Contrast: WebGLUniformLocation;
-    Gamma: WebGLUniformLocation;
-    Alpha: WebGLUniformLocation;
-    ScaleType: WebGLUniformLocation;
-    Inverted: WebGLUniformLocation;
-    NaNColor: WebGLUniformLocation;
-    DataTexture: WebGLUniformLocation;
-    CmapTexture: WebGLUniformLocation;
-    NumCmaps: WebGLUniformLocation;
-    CmapIndex: WebGLUniformLocation;
-    CanvasWidth: WebGLUniformLocation;
-    CanvasHeight: WebGLUniformLocation;
-    RotationOrigin: WebGLUniformLocation;
-    RotationAngle: WebGLUniformLocation;
-    ScaleAdjustment: WebGLUniformLocation;
-    TiledRendering: WebGLUniformLocation;
-    TileSize: WebGLUniformLocation;
-    TileScaling: WebGLUniformLocation;
-    TileOffset: WebGLUniformLocation;
-    TileTextureOffset: WebGLUniformLocation;
-    TileTextureSize: WebGLUniformLocation;
-    TextureSize: WebGLUniformLocation;
-    TileBorder: WebGLUniformLocation;
-}
-
 @observer
 export class RasterViewComponent extends React.Component<RasterViewComponentProps> {
     private canvas: HTMLCanvasElement;
     private gl: WebGLRenderingContext;
-    private hasFloatExtension: boolean;
-    // GL buffers
-    private rasterDataBuffer: ArrayBufferLike;
-    private cmapTexture: WebGLTexture;
-    private animationDataTexture: WebGLTexture;
-    private vertexPositionBuffer: WebGLBuffer;
-    private animationVertexPositionBuffer: WebGLBuffer;
-    private vertexUVBuffer: WebGLBuffer;
-    // Shader attribute handles
-    private vertexPositionAttribute: number;
-    private vertexUVAttribute: number;
-    // Shader uniform handles
-    private shaderProgram: WebGLProgram;
-    private shaderUniforms: ShaderUniforms;
 
     componentDidMount() {
+        this.gl = this.props.tileService.tileWebGLService.gl;
         if (this.canvas) {
-            this.initWebGL();
-            this.initShaders();
-            this.initBuffers();
-            loadImageTexture(this.gl, allMaps, WebGLRenderingContext.TEXTURE1).then(texture => {
-                this.cmapTexture = texture;
-                this.updateCanvas();
-            });
+            this.updateCanvas();
         }
-        this.props.tileService.GetTileStream().subscribe(tileCount => {
-            requestAnimationFrame(this.updateCanvas);
-        });
-    }
-
-    componentWillUnmount() {
-        console.log(`Unmounting image view, cleaning up WebGL resources`);
-        // Attempt to clean up WebGL resources
-        const numTextureUnits = this.gl.getParameter(WebGLRenderingContext.MAX_TEXTURE_IMAGE_UNITS);
-        for (let unit = 0; unit < numTextureUnits; ++unit) {
-            this.gl.activeTexture(WebGLRenderingContext.TEXTURE0 + unit);
-            this.gl.bindTexture(WebGLRenderingContext.TEXTURE_2D, null);
-            this.gl.bindTexture(WebGLRenderingContext.TEXTURE_CUBE_MAP, null);
-        }
-        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, null);
-        this.gl.bindBuffer(WebGLRenderingContext.ELEMENT_ARRAY_BUFFER, null);
-        this.gl.bindRenderbuffer(WebGLRenderingContext.RENDERBUFFER, null);
-        this.gl.bindFramebuffer(WebGLRenderingContext.FRAMEBUFFER, null);
-
-        if (this.props.tileService) {
-            this.props.tileService.clearContext();
-        }
-
-        this.gl.deleteTexture(this.cmapTexture);
-        this.gl.deleteTexture(this.animationDataTexture);
-        this.gl.deleteBuffer(this.vertexPositionBuffer);
-        this.gl.deleteBuffer(this.animationVertexPositionBuffer);
-        this.gl.deleteBuffer(this.vertexUVBuffer);
-        this.gl.deleteProgram(this.shaderProgram);
-
-        this.gl.canvas.width = 1;
-        this.gl.canvas.height = 1;
-        this.gl.getExtension("WEBGL_lose_context").loseContext();
+        this.props.tileService.tileStream.subscribe(() => requestAnimationFrame(this.updateCanvas));
     }
 
     componentDidUpdate() {
         requestAnimationFrame(this.updateCanvas);
     }
 
-    private initWebGL() {
-        try {
-            this.gl = this.canvas.getContext("webgl", {preserveDrawingBuffer: true});
-            if (!this.gl) {
-                return;
-            }
-        } catch (e) {
-            console.log(e);
-        }
-        if (!this.gl) {
-            console.log("Could not initialise WebGL");
-        }
-
-        const extTextureFloat = this.gl.getExtension("OES_texture_float");
-
-        if (!extTextureFloat) {
-            console.log("Could not initialise WebGL float texture extension");
-            this.hasFloatExtension = false;
-        } else {
-            this.hasFloatExtension = true;
-        }
-        this.props.tileService.setContext(this.gl);
-    }
-
     private updateCanvas = () => {
         const frame = this.props.frame;
-        if (frame && this.canvas && this.gl && this.cmapTexture) {
-            if (frame.renderType === RasterRenderType.ANIMATION && frame.rasterData && frame.rasterData.buffer !== this.rasterDataBuffer) {
+        const tileRenderService = this.props.tileService.tileWebGLService;
+        if (frame && this.canvas && this.gl && tileRenderService.cmapTexture) {
+            if (frame.renderType === RasterRenderType.ANIMATION && frame.rasterData && frame.rasterData.buffer !== tileRenderService.rasterDataBuffer) {
                 this.updateTexture();
             }
 
             this.updateCanvasSize();
             this.updateUniforms();
             this.renderCanvas();
+            // draw in 2d canvas
+            const ctx = this.canvas.getContext("2d");
+            ctx.clearRect(0, 0, this.canvas.width, this.canvas.height);
+            ctx.drawImage(this.gl.canvas, 0, 0, this.canvas.width, this.canvas.height);
         }
     };
 
     private updateTexture() {
+        const tileRenderService = this.props.tileService.tileWebGLService;
         const frame = this.props.frame;
         const w = Math.ceil((frame.currentFrameView.xMax - frame.currentFrameView.xMin) / frame.currentFrameView.mip);
         const h = Math.ceil((frame.currentFrameView.yMax - frame.currentFrameView.yMin) / frame.currentFrameView.mip);
@@ -160,42 +58,46 @@ export class RasterViewComponent extends React.Component<RasterViewComponentProp
             console.log(`Data mismatch! L=${frame.rasterData ? frame.rasterData.length : "null"}, WxH = ${w}x${h}=${w * h}`);
             return;
         } else {
-            this.gl.bindTexture(WebGLRenderingContext.TEXTURE_2D, this.animationDataTexture);
+            this.gl.bindTexture(WebGLRenderingContext.TEXTURE_2D, tileRenderService.animationDataTexture);
             loadFP32Texture(this.gl, frame.rasterData, w, h, w, h, WebGLRenderingContext.TEXTURE0);
-            this.rasterDataBuffer = frame.rasterData.buffer;
+            tileRenderService.rasterDataBuffer = frame.rasterData.buffer;
         }
     }
 
     private updateUniforms() {
+        const tileRenderService = this.props.tileService.tileWebGLService;
         const frame = this.props.frame;
         const renderConfig = frame.renderConfig;
         const preference = this.props.preference;
-        if (renderConfig && preference && this.shaderUniforms) {
-            this.gl.uniform1f(this.shaderUniforms.MinVal, renderConfig.scaleMinVal);
-            this.gl.uniform1f(this.shaderUniforms.MaxVal, renderConfig.scaleMaxVal);
-            this.gl.uniform1i(this.shaderUniforms.CmapIndex, renderConfig.colorMap);
-            this.gl.uniform1i(this.shaderUniforms.ScaleType, renderConfig.scaling);
-            this.gl.uniform1i(this.shaderUniforms.Inverted, renderConfig.inverted ? 1 : 0);
-            this.gl.uniform1f(this.shaderUniforms.Bias, renderConfig.bias);
-            this.gl.uniform1f(this.shaderUniforms.Contrast, renderConfig.contrast);
-            this.gl.uniform1f(this.shaderUniforms.Gamma, renderConfig.gamma);
-            this.gl.uniform1f(this.shaderUniforms.Alpha, renderConfig.alpha);
-            this.gl.uniform1f(this.shaderUniforms.CanvasWidth, frame.renderWidth * devicePixelRatio);
-            this.gl.uniform1f(this.shaderUniforms.CanvasHeight, frame.renderHeight * devicePixelRatio);
+        if (renderConfig && preference && tileRenderService.shaderUniforms) {
+            this.gl.uniform1f(tileRenderService.shaderUniforms.MinVal, renderConfig.scaleMinVal);
+            this.gl.uniform1f(tileRenderService.shaderUniforms.MaxVal, renderConfig.scaleMaxVal);
+            this.gl.uniform1i(tileRenderService.shaderUniforms.CmapIndex, renderConfig.colorMap);
+            this.gl.uniform1i(tileRenderService.shaderUniforms.ScaleType, renderConfig.scaling);
+            this.gl.uniform1i(tileRenderService.shaderUniforms.Inverted, renderConfig.inverted ? 1 : 0);
+            this.gl.uniform1f(tileRenderService.shaderUniforms.Bias, renderConfig.bias);
+            this.gl.uniform1f(tileRenderService.shaderUniforms.Contrast, renderConfig.contrast);
+            this.gl.uniform1f(tileRenderService.shaderUniforms.Gamma, renderConfig.gamma);
+            this.gl.uniform1f(tileRenderService.shaderUniforms.Alpha, renderConfig.alpha);
+            this.gl.uniform1f(tileRenderService.shaderUniforms.CanvasWidth, frame.renderWidth * devicePixelRatio);
+            this.gl.uniform1f(tileRenderService.shaderUniforms.CanvasHeight, frame.renderHeight * devicePixelRatio);
 
             const rgba = hexStringToRgba(preference.nanColorHex, preference.nanAlpha);
             if (rgba) {
-                this.gl.uniform4f(this.shaderUniforms.NaNColor, rgba.r / 255, rgba.g / 255, rgba.b / 255, rgba.a);
+                this.gl.uniform4f(tileRenderService.shaderUniforms.NaNColor, rgba.r / 255, rgba.g / 255, rgba.b / 255, rgba.a);
             }
         }
     }
 
     private updateCanvasSize() {
         const frame = this.props.frame;
+        const tileRenderService = this.props.tileService.tileWebGLService;
+
         // Resize and clear the canvas if needed
         if (frame && frame.isRenderable && (this.canvas.width !== frame.renderWidth * devicePixelRatio || this.canvas.height !== frame.renderHeight * devicePixelRatio)) {
             this.canvas.width = frame.renderWidth * devicePixelRatio;
             this.canvas.height = frame.renderHeight * devicePixelRatio;
+            tileRenderService.setCanvasSize(frame.renderWidth * devicePixelRatio, frame.renderHeight * devicePixelRatio);
         }
     }
 
@@ -221,6 +123,7 @@ export class RasterViewComponent extends React.Component<RasterViewComponentProp
 
     private renderAnimationCanvas() {
         const frame = this.props.frame;
+        const tileRenderService = this.props.tileService.tileWebGLService;
 
         const current = frame.currentFrameView;
         const full = frame.requiredFrameView;
@@ -243,24 +146,25 @@ export class RasterViewComponent extends React.Component<RasterViewComponentProp
         ].map(v => -1 + 2 * v));
 
         this.gl.activeTexture(WebGLRenderingContext.TEXTURE0);
-        this.gl.bindTexture(WebGLRenderingContext.TEXTURE_2D, this.animationDataTexture);
-        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, this.vertexUVBuffer);
-        this.gl.vertexAttribPointer(this.vertexUVAttribute, 2, WebGLRenderingContext.FLOAT, false, 0, 0);
-        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, this.animationVertexPositionBuffer);
-        this.gl.vertexAttribPointer(this.vertexPositionAttribute, 3, WebGLRenderingContext.FLOAT, false, 0, 0);
+        this.gl.bindTexture(WebGLRenderingContext.TEXTURE_2D, tileRenderService.animationDataTexture);
+        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, tileRenderService.vertexUVBuffer);
+        this.gl.vertexAttribPointer(tileRenderService.vertexUVAttribute, 2, WebGLRenderingContext.FLOAT, false, 0, 0);
+        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, tileRenderService.animationVertexPositionBuffer);
+        this.gl.vertexAttribPointer(tileRenderService.vertexPositionAttribute, 3, WebGLRenderingContext.FLOAT, false, 0, 0);
         this.gl.bufferData(WebGLRenderingContext.ARRAY_BUFFER, new Float32Array(vertices), WebGLRenderingContext.DYNAMIC_DRAW);
-        this.gl.uniform1i(this.shaderUniforms.TiledRendering, 0);
+        this.gl.uniform1i(tileRenderService.shaderUniforms.TiledRendering, 0);
         this.gl.drawArrays(WebGLRenderingContext.TRIANGLE_STRIP, 0, 4);
     }
 
     private renderTiledCanvas() {
         const frame = this.props.frame;
+        const tileRenderService = this.props.tileService.tileWebGLService;
 
-        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, this.vertexUVBuffer);
-        this.gl.vertexAttribPointer(this.vertexUVAttribute, 2, WebGLRenderingContext.FLOAT, false, 0, 0);
-        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, this.vertexPositionBuffer);
-        this.gl.vertexAttribPointer(this.vertexPositionAttribute, 3, WebGLRenderingContext.FLOAT, false, 0, 0);
-        this.gl.uniform1i(this.shaderUniforms.TiledRendering, 1);
+        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, tileRenderService.vertexUVBuffer);
+        this.gl.vertexAttribPointer(tileRenderService.vertexUVAttribute, 2, WebGLRenderingContext.FLOAT, false, 0, 0);
+        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, tileRenderService.vertexPositionBuffer);
+        this.gl.vertexAttribPointer(tileRenderService.vertexPositionAttribute, 3, WebGLRenderingContext.FLOAT, false, 0, 0);
+        this.gl.uniform1i(tileRenderService.shaderUniforms.TiledRendering, 1);
 
         const imageSize = {x: frame.frameInfo.fileInfoExtended.width, y: frame.frameInfo.fileInfoExtended.height};
         const boundedView: FrameView = {
@@ -350,6 +254,7 @@ export class RasterViewComponent extends React.Component<RasterViewComponentProp
 
     private renderTile(tile: TileCoordinate, rasterTile: RasterTile, mip: number) {
         const frame = this.props.frame;
+        const tileRenderService = this.props.tileService.tileWebGLService;
 
         if (!rasterTile) {
             return;
@@ -365,7 +270,7 @@ export class RasterViewComponent extends React.Component<RasterViewComponentProp
             this.gl.bindTexture(WebGLRenderingContext.TEXTURE_2D, textureParameters.texture);
             this.gl.texParameteri(WebGLRenderingContext.TEXTURE_2D, WebGLRenderingContext.TEXTURE_MIN_FILTER, GL.NEAREST);
             this.gl.texParameteri(WebGLRenderingContext.TEXTURE_2D, WebGLRenderingContext.TEXTURE_MAG_FILTER, GL.NEAREST);
-            this.gl.uniform2f(this.shaderUniforms.TileTextureOffset, textureParameters.offset.x, textureParameters.offset.y);
+            this.gl.uniform2f(tileRenderService.shaderUniforms.TileTextureOffset, textureParameters.offset.x, textureParameters.offset.y);
         }
 
         const spatialRef = frame.spatialReference || frame;
@@ -392,108 +297,21 @@ export class RasterViewComponent extends React.Component<RasterViewComponentProp
                 x: spatialRef.zoomLevel * (rotationOriginImageSpace.x - full.xMin),
                 y: spatialRef.zoomLevel * (rotationOriginImageSpace.y - full.yMin),
             };
-            this.gl.uniform2f(this.shaderUniforms.RotationOrigin, rotationOriginCanvasSpace.x, rotationOriginCanvasSpace.y);
-            this.gl.uniform1f(this.shaderUniforms.RotationAngle, -frame.spatialTransform.rotation);
-            this.gl.uniform1f(this.shaderUniforms.ScaleAdjustment, frame.spatialTransform.scale);
+            this.gl.uniform2f(tileRenderService.shaderUniforms.RotationOrigin, rotationOriginCanvasSpace.x, rotationOriginCanvasSpace.y);
+            this.gl.uniform1f(tileRenderService.shaderUniforms.RotationAngle, -frame.spatialTransform.rotation);
+            this.gl.uniform1f(tileRenderService.shaderUniforms.ScaleAdjustment, frame.spatialTransform.scale);
         } else {
-            this.gl.uniform1f(this.shaderUniforms.RotationAngle, 0);
-            this.gl.uniform1f(this.shaderUniforms.ScaleAdjustment, 1);
+            this.gl.uniform1f(tileRenderService.shaderUniforms.RotationAngle, 0);
+            this.gl.uniform1f(tileRenderService.shaderUniforms.ScaleAdjustment, 1);
         }
 
         // take zoom level into account to convert from image space to canvas space
         bottomLeft = scale2D(bottomLeft, spatialRef.zoomLevel);
 
-        this.gl.uniform2f(this.shaderUniforms.TileSize, rasterTile.width, rasterTile.height);
-        this.gl.uniform2f(this.shaderUniforms.TileOffset, bottomLeft.x, bottomLeft.y);
-        this.gl.uniform2f(this.shaderUniforms.TileScaling, tileScaling.x, tileScaling.y);
+        this.gl.uniform2f(tileRenderService.shaderUniforms.TileSize, rasterTile.width, rasterTile.height);
+        this.gl.uniform2f(tileRenderService.shaderUniforms.TileOffset, bottomLeft.x, bottomLeft.y);
+        this.gl.uniform2f(tileRenderService.shaderUniforms.TileScaling, tileScaling.x, tileScaling.y);
         this.gl.drawArrays(WebGLRenderingContext.TRIANGLE_STRIP, 0, 4);
-    }
-
-    private initShaders() {
-        this.shaderProgram = getShaderProgram(this.gl, vertexShader, pixelShader);
-        this.gl.useProgram(this.shaderProgram);
-
-        this.vertexPositionAttribute = this.gl.getAttribLocation(this.shaderProgram, "aVertexPosition");
-        this.vertexUVAttribute = this.gl.getAttribLocation(this.shaderProgram, "aVertexUV");
-        this.gl.enableVertexAttribArray(this.vertexPositionAttribute);
-        this.gl.enableVertexAttribArray(this.vertexUVAttribute);
-
-        this.shaderUniforms = {
-            MinVal: this.gl.getUniformLocation(this.shaderProgram, "uMinVal"),
-            MaxVal: this.gl.getUniformLocation(this.shaderProgram, "uMaxVal"),
-            NaNColor: this.gl.getUniformLocation(this.shaderProgram, "uNaNColor"),
-            Bias: this.gl.getUniformLocation(this.shaderProgram, "uBias"),
-            Contrast: this.gl.getUniformLocation(this.shaderProgram, "uContrast"),
-            Gamma: this.gl.getUniformLocation(this.shaderProgram, "uGamma"),
-            Alpha: this.gl.getUniformLocation(this.shaderProgram, "uAlpha"),
-            ScaleType: this.gl.getUniformLocation(this.shaderProgram, "uScaleType"),
-            Inverted: this.gl.getUniformLocation(this.shaderProgram, "uInverted"),
-            DataTexture: this.gl.getUniformLocation(this.shaderProgram, "uDataTexture"),
-            CmapTexture: this.gl.getUniformLocation(this.shaderProgram, "uCmapTexture"),
-            NumCmaps: this.gl.getUniformLocation(this.shaderProgram, "uNumCmaps"),
-            CmapIndex: this.gl.getUniformLocation(this.shaderProgram, "uCmapIndex"),
-            CanvasWidth: this.gl.getUniformLocation(this.shaderProgram, "uCanvasWidth"),
-            CanvasHeight: this.gl.getUniformLocation(this.shaderProgram, "uCanvasHeight"),
-            ScaleAdjustment: this.gl.getUniformLocation(this.shaderProgram, "uScaleAdjustment"),
-            RotationOrigin: this.gl.getUniformLocation(this.shaderProgram, "uRotationOrigin"),
-            RotationAngle: this.gl.getUniformLocation(this.shaderProgram, "uRotationAngle"),
-            TiledRendering: this.gl.getUniformLocation(this.shaderProgram, "uTiledRendering"),
-            TileSize: this.gl.getUniformLocation(this.shaderProgram, "uTileSize"),
-            TileScaling: this.gl.getUniformLocation(this.shaderProgram, "uTileScaling"),
-            TileOffset: this.gl.getUniformLocation(this.shaderProgram, "uTileOffset"),
-            TileTextureOffset: this.gl.getUniformLocation(this.shaderProgram, "uTileTextureOffset"),
-            TextureSize: this.gl.getUniformLocation(this.shaderProgram, "uTextureSize"),
-            TileTextureSize: this.gl.getUniformLocation(this.shaderProgram, "uTileTextureSize"),
-            TileBorder: this.gl.getUniformLocation(this.shaderProgram, "uTileBorder")
-        };
-
-        this.gl.uniform1i(this.shaderUniforms.DataTexture, 0);
-        this.gl.uniform1i(this.shaderUniforms.CmapTexture, 1);
-        this.gl.uniform1i(this.shaderUniforms.NumCmaps, 79);
-        this.gl.uniform1i(this.shaderUniforms.CmapIndex, 2);
-        this.gl.uniform1f(this.shaderUniforms.MinVal, 3.4);
-        this.gl.uniform1f(this.shaderUniforms.MaxVal, 5.50);
-        this.gl.uniform1f(this.shaderUniforms.Bias, 0);
-        this.gl.uniform1f(this.shaderUniforms.Contrast, 1);
-        this.gl.uniform1f(this.shaderUniforms.Gamma, 1);
-        this.gl.uniform1f(this.shaderUniforms.Alpha, 1000);
-        this.gl.uniform1i(this.shaderUniforms.Inverted, 0);
-        this.gl.uniform1i(this.shaderUniforms.TiledRendering, 1);
-        this.gl.uniform1f(this.shaderUniforms.TileBorder, 0 / TILE_SIZE);
-        this.gl.uniform2f(this.shaderUniforms.TileSize, 1, 1);
-        this.gl.uniform2f(this.shaderUniforms.TileScaling, 1, 1);
-        this.gl.uniform2f(this.shaderUniforms.TileOffset, 0, 0);
-        this.gl.uniform2f(this.shaderUniforms.TileTextureOffset, 0, 0);
-        this.gl.uniform1f(this.shaderUniforms.TextureSize, TEXTURE_SIZE);
-        this.gl.uniform1f(this.shaderUniforms.TileTextureSize, TILE_SIZE);
-        this.gl.uniform4f(this.shaderUniforms.NaNColor, 0, 0, 1, 1);
-    }
-
-    private initBuffers() {
-        this.vertexPositionBuffer = this.gl.createBuffer();
-        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, this.vertexPositionBuffer);
-        const vertices = new Float32Array([
-            0.0, 0.0, 0,
-            1.0, 0.0, 0,
-            0.0, 1.0, 0,
-            1.0, 1.0, 0
-        ]);
-        this.gl.bufferData(WebGLRenderingContext.ARRAY_BUFFER, vertices, WebGLRenderingContext.STATIC_DRAW);
-
-        this.vertexUVBuffer = this.gl.createBuffer();
-        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, this.vertexUVBuffer);
-        const uvs = new Float32Array([
-            0.0, 0.0,
-            1.0, 0.0,
-            0.0, 1.0,
-            1.0, 1.0
-        ]);
-        this.gl.bufferData(WebGLRenderingContext.ARRAY_BUFFER, uvs, WebGLRenderingContext.STATIC_DRAW);
-
-        this.animationVertexPositionBuffer = this.gl.createBuffer();
-        this.gl.bindBuffer(WebGLRenderingContext.ARRAY_BUFFER, this.animationVertexPositionBuffer);
-        this.gl.bufferData(WebGLRenderingContext.ARRAY_BUFFER, vertices, WebGLRenderingContext.DYNAMIC_DRAW);
-        this.animationDataTexture = this.gl.createTexture();
     }
 
     render() {
