+++ src/services/GLSL/pixel_shader_float_rgb.glsl
