+++ src/services/index.ts
@@ -1,3 +1,4 @@
 export * from "./BackendService";
 export * from "./DecompressionService";
 export * from "./TileService";
+export * from "./TileWebGLService";
